requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP客户端
httpx==0.25.2

# 原有Flask依赖（保留兼容性）
flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# 文件处理
python-dotenv==1.0.0

# 原项目其他依赖
openai
langchain
langchain-community
langchain-core
huggingface_hub
sentence-transformers